seperate stockx scraper from the depop one
