+++ fashioncrawler/scraper/__init__.py
@@ -1,3 +1,4 @@
 from .base_scraper import BaseScraper
 from .grailed_scraper import GrailedScraper
 from .depop_scraper import DepopScraper
+from .stockx_scraper import StockxScraper
