+++ fashioncrawler/scraper/depop_scraper.py
@@ -327,20 +327,3 @@ class DepopScraper(BaseScraper):
                 time.sleep(backoff_delay * retries)  # Increasing delay for retries
             finally:
                 driver.quit()
-
-
-class StockxScraper(BaseScraper):
-    """
-    Subclass of BaseScraper for scraping data from the StockX website.
-    """
-
-    def __init__(self, base_scraper):
-        self.driver = base_scraper.driver
-
-    def get_to_search_bar_to_search(
-        self, search_bar_css_selector: str, timeout=2
-    ) -> None:
-        pass
-
-    def run_scraper(self, search_query):
-        pass
